Update thermo_integ.cpp
In this version the delta_p and delta_q values are calculated based on the p_initial, p_final, q_intial and q_final values. 
Also, the epsilon[i][i] are first obtained and then the values of epsilon[i][i] for typeA is modified and finally epsilon[i][j] for affected pairs is modified.